requirements.txt
Django==2.2.4
djangorestframework==3.10.2
GitPython==3.0.2
python-dateutil==2.8.0
pytz==2019.2
